Added very simple linewrapping for the legends, and naming in the tips.
Should it be a max character limit in the name field of the plot modal? This would make it easier to avoid overlapping. On the other hand, we could leave it up to the user to avoid overlapping text.

Also, this is what I meant with the name in the tips. Is this sufficient?

Off-topic: I consider trying to add sliders to modify custom variables, to be placed under the function-plots. Is this something of interest? I will probably try to make it for myself either way..